+++ host.go
@@ -47,14 +47,14 @@ type HostStatsData struct {
 
 // SearchOptions options of search, for post processors
 type SearchOptions struct {
-	FixUrl     bool   // each host fix as url, like 1.1.1.1,80 will change to http://1.1.1.1, https://1.1.1.1:8443 will no change
-	UrlPrefix  string // default is http://
-	Full       bool   // search result for over a year
-	UniqByIP   bool   // uniq by ip
-	IsActive   bool   // probe website is existed, add isActive field
-	DeWildcard int    // number of wildcard domains retained
-	Filter     string // filter data by rules
-	DedupHost  bool   // prioritize subdomain data retention
+	FixUrl      bool   // each host fix as url, like 1.1.1.1,80 will change to http://1.1.1.1, https://1.1.1.1:8443 will no change
+	UrlPrefix   string // default is http://
+	Full        bool   // search result for over a year
+	UniqByIP    bool   // uniq by ip
+	CheckActive int    // probe website is existed, add isActive field
+	DeWildcard  int    // number of wildcard domains retained
+	Filter      string // filter data by rules
+	DedupHost   bool   // prioritize subdomain data retention
 }
 
 // fixHostToUrl 替换host为url
@@ -184,7 +184,7 @@ func (c *Client) HostSearch(query string, size int, fields []string, options ...
 	var (
 		full        bool
 		uniqByIP    bool
-		isActive    bool
+		isActive    int
 		dedupCname  int
 		isSubDomain bool
 		filter      string
@@ -192,7 +192,7 @@ func (c *Client) HostSearch(query string, size int, fields []string, options ...
 	if len(options) > 0 {
 		full = options[0].Full
 		uniqByIP = options[0].UniqByIP
-		isActive = options[0].IsActive
+		isActive = options[0].CheckActive
 		dedupCname = options[0].DeWildcard
 		filter = options[0].Filter
 		isSubDomain = options[0].DedupHost
@@ -247,9 +247,10 @@ func (c *Client) HostSearch(query string, size int, fields []string, options ...
 
 	var activeSlice []string
 	// 确认fields包含link
-	var linkIndex = -1
-	if isActive {
+	var linkIndex, codeIndex = -1, -1
+	if isActive > 0 {
 		linkIndex, fields = getParamIndexThenAdd(fields, "link")
+		codeIndex, fields = getParamIndexThenAdd(fields, "status_code")
 	}
 
 	dedupCnameMap := make(map[string]int)
@@ -365,9 +366,10 @@ func (c *Client) HostSearch(query string, size int, fields []string, options ...
 							continue
 						}
 					}
-					if isActive {
-						active := CheckActive(newSlice[linkIndex])
-						activeSlice = append(activeSlice, fmt.Sprintf("%t", active))
+					if isActive > 0 {
+						resp := DoHttpCheck(newSlice[linkIndex], isActive)
+						activeSlice = append(activeSlice, fmt.Sprintf("%t", resp.IsActive))
+						newSlice[codeIndex] = resp.StatusCode
 					}
 					results = append(results, newSlice)
 				} else if vStr, ok := result.(string); ok {
@@ -379,9 +381,9 @@ func (c *Client) HostSearch(query string, size int, fields []string, options ...
 						}
 						uniqIPMap[vStr] = true
 					}
-					if isActive && linkIndex == 0 {
-						active := CheckActive(vStr)
-						activeSlice = append(activeSlice, fmt.Sprintf("%t", active))
+					if isActive > 0 && linkIndex == 0 {
+						resp := DoHttpCheck(vStr, isActive)
+						activeSlice = append(activeSlice, fmt.Sprintf("%t", resp.IsActive))
 					}
 					results = append(results, newSlice)
 				}
@@ -434,7 +436,7 @@ func (c *Client) HostSearch(query string, size int, fields []string, options ...
 
 	// 后处理
 	res = c.postProcess(res, fields, hostIndex, protocolIndex, rawFieldSize, options...)
-	if isActive {
+	if isActive > 0 {
 		for index := range res {
 			res[index] = append(res[index], activeSlice[index])
 		}